docs: :sparkles: update index and remove deprecated API examples
This commit updates the documentation by modifying the index page to replace the "Contribution Guide" link with a new "Components" link that directs to the "Data Tables" guide. Additionally, the outdated "API Examples" page has been removed to streamline the documentation and focus on more relevant content. These changes enhance the clarity and usability of the documentation for developers.